[/] [Code] Args are pointers to struc..
All args like (TMemorySegment Src) are now (TMemorySegment * Src).
Or else the structure is copied. Not a big deal until you have a class
with managed memory and destructor.

Also GetByte() is moved under TMemorySegment. Yes, this is trivial
and it should have trivial SetByte() counterpart. Things may become
more useful if I add GetWord() and GetLong().. Or GetUint_4()?

That's not cleat for me, and not needed now. I would likely extract
that getters/setters to separate file when they'll grow large enough.