[/] [Code] CopyMem() arguments are again plain structs, not pointers..
because it's natural. When I'm passing struct in argument it means
I want to pass group of related arguments. Not the pointer to group.

Because C will start worrying about what will happen if code will
modify data at that pointer. And will nudge to use "const" attribute.
That's road to madness.

Passing immediate structure works as long as structure does not have
side-effects destructor. When it does, passing pointer is only option.