[+] [Interface] "As progmem seg" macro..
Well, I really tried to avoid adding macro to this codebase.

But we want more SRAM for larger projects. We can get it by storing
strings in program (in program memory, flash).

Compare two snippets:

  printf_P(PSTR("Whoah!! Arduino!!!\n"));

and

  static const TUint_1 ProgmemData[] [[gnu::progmem]] =
    "This is how to print data from program memory";

  me_MemorySegment::TMemorySegment ProgmemSeg =
    {
      { .Addr = (TAddress) ProgmemData },
      .Size = sizeof(ProgmemData) - 1
    };

  Console.PrintProgmem(ProgmemSeg);

We're losing in usability.

And also naming every asciiz chunk is silly.

So now we have macro "AsProgmemSeg". It expands ASCIIZ const to that
code.

Usage

  Console.PrintProgmem(AsProgmemSeg("This is how to print data from program memory"));