[!^-] [Interface] WHEEE! We removed <.Bytes> from segment definition..
So yes, you're no longer can write "MemSeg.Bytes[5] = 42".

    union
    {
      TAddress Addr = 0;
      // "Bytes" provides array access to memory at "Addr"
      TUnit * Bytes;
    };

Was hack we considered cool and sometimes useful to write data
comparator and copier. Later we discovered that there more memories
and our comparator and copier wont work for them.

So we wrote accessors for those memories. And we wrote streams to
use accessors. And we moved comparator and copier to streams.

It's not good for code and memory footprint. More indirections,
GCC can no longer embed direct memory instructions instead of
function calls.

But we're more satisfied with code. More readable and more reliable
by design.